% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_axis.R
\name{central_axis}
\alias{central_axis}
\title{Define muscle force axis vector}
\usage{
central_axis(origin_centroid, insertion_centroid)
}
\arguments{
\item{origin_centroid}{numeric vector of origin centroid, or equivalent
point}

\item{insertion_centroid}{numeric vector of insertion centroid, or
equivalent point}
}
\value{
Named numeric vector of the axis between the origin and insertion
centroids
}
\description{
Calculate the force axis for a given muscle using two points.
Points are assumed to be the centroids of origin and insertion,
but can be other points as needed.
}
\examples{

central_axis(c(0, 0, 0), c(1, 1, 1))

}
