% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_location.R
\name{centroid_location}
\alias{centroid_location}
\title{Calculate centroid location}
\usage{
centroid_location(fname)
}
\arguments{
\item{fname}{String: Path to stl file}
}
\value{
Numeric vector of centroid location
}
\description{
Calculate centroid location
}
\examples{
f <- system.file("extdata", "L_mPTd_Or.stl",
                 package = "MuscleTernary")
centroid_location(f)
}
