% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_size.R
\name{centroid_size}
\alias{centroid_size}
\title{Calculate centroid size for an stl file}
\usage{
centroid_size(fname)
}
\arguments{
\item{fname}{String: Path to stl file}
}
\value{
Numeric: Centroid size
}
\description{
Calculate centroid size for an stl file
}
\examples{
f <- system.file("extdata", "L_mPTd_Or.stl",
                 package = "MuscleTernary")
centroid_size(f)
}
