% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_process.R
\name{coords_process}
\alias{coords_process}
\title{Process coordinates and force files.}
\usage{
coords_process(coords_infile, force_info)
}
\arguments{
\item{coords_infile}{String giving the path to a coordinates file.}

\item{force_info}{String giving the path to the force information
file.}
}
\value{
`data.frame` with coordinate and forces files combined.
}
\description{
Process coordinates and force files.
}
