% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CrossProduct3D}
\alias{CrossProduct3D}
\title{Compute 3D vector cross product}
\usage{
CrossProduct3D(x, y, i = 1:3)
}
\arguments{
\item{x}{Vector or matrix}

\item{y}{Vector or matrix}

\item{i}{Indices}
}
\value{
Cross product matrix
}
\description{
Compute the vector cross product between x and y, and return the
components indexed by i.
See: https://stackoverflow.com/a/21736807/168137
}
\keyword{internal}
