% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-MuscleTernary.R
\name{dot}
\alias{dot}
\title{Dot product}
\usage{
dot(u, v)
}
\arguments{
\item{u}{Numeric vector}

\item{v}{Numeric vector}
}
\value{
Numeric dot product of `u` and `v`.
}
\description{
Dot product
}
\examples{
dot(c(1, 0, 0), c(0, 1, 0))
dot(c(1, 2, 3), c(4, 5, 6))

}
\keyword{internal}
