% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{euler}
\alias{euler}
\title{Decompose rotation matrix U into Euler angles}
\usage{
euler(U)
}
\arguments{
\item{U}{Rotation matrix}
}
\value{
Euler angles
}
\description{
Note that the order
is x, y, x.
See: http://nghiaho.com/?page_id=846
}
\keyword{internal}
