% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_euler_angles}
\alias{get_euler_angles}
\title{Calculate the Euler angles for rotating A onto B.}
\usage{
get_euler_angles(A, B)
}
\arguments{
\item{A}{Vector (usually the unit vector)}

\item{B}{Vector}
}
\value{
Euler angles between A and B
}
\description{
The utility is to set up geometry with a known normal (0, 1, 0),
and calculate the xyz rotation sequence to match another vector.
}
