% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_xyz.R
\name{interpolate_xyz}
\alias{interpolate_xyz}
\title{Interpolate (x, y, z) coordinates}
\usage{
interpolate_xyz(start, end, length_out)
}
\arguments{
\item{start}{Vector of xyz coordinate for start position.}

\item{end}{Vector of xyz coordinate for end position.}

\item{length_out}{Total length of output `data.frame`.}
}
\value{
`data.frame` interpolated between `start` and `end` with
  column for frame number.
}
\description{
Given a start and end point for a pair of xyz coordinate, return
the linearly interpolated data.frame between them. This functional
also adds a `.frame` column that allows animation.
}
