% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag.R
\name{mag}
\alias{mag}
\title{Calculate vector magnitude}
\usage{
mag(vector)
}
\arguments{
\item{vector}{Vector whose magnitude you want to measure}
}
\value{
Numeric: Magnitude of the vector (x, y, z).
}
\description{
Calculates the magnitude of a given vector
}
\keyword{internal}
