% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mel.R
\name{make_mel}
\alias{make_mel}
\title{Write mel file}
\usage{
make_mel(
  stl,
  data,
  shader_file = "default",
  outfile = NULL,
  scale_radius = TRUE,
  use_stl = FALSE,
  max_radius = 8,
  rev_arrows = TRUE,
  write_file = TRUE
)
}
\arguments{
\item{stl}{string Location of the stl file. Assumed to present in the
working directory}

\item{data}{data.frame Object with the data}

\item{shader_file}{string Either "default" to use the standard palette,
or the path to a csv file with colors for each muscle.}

\item{outfile}{string Name of file to write to. Defaults to the base name
of the stl file.}

\item{scale_radius}{boolean (default \code{TRUE}) Should the radius be
scaled based on the force variable?}

\item{use_stl}{boolean (default \code{FALSE}. Should scaling use the
maximum value automatically based on the stl centroid size?}

\item{max_radius}{numeric Maximum radius value. Ignored if
\code{scale_radius} is \code{TRUE}.}

\item{rev_arrows}{boolean (default \code{TRUE}) Should the arrowheads be
reversed (you probably want \code{TRUE})}

\item{write_file}{boolean (default \code{TRUE}) Should the mel file be
written out}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of
  writing a Maya mel script file.
}
\description{
Write mel file
}
\examples{
\dontrun{
data_file <- system.file("extdata", "AL_008_data.csv",
                         package = "MuscleTernary")
dat <- readr::read_csv(data_file)
make_mel("L_mPTd_Or.stl", dat)
}
}
