% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unit_vector.R
\name{make_unit_vector}
\alias{make_unit_vector}
\title{Create a unit vector from a given vector}
\usage{
make_unit_vector(x)
}
\arguments{
\item{x}{Vector of length 3.}
}
\value{
A normalized unit vector of length 3
}
\description{
Create a unit vector from a given vector
}
\examples{
make_unit_vector(c(3, 4, 0))
make_unit_vector(c(1, 1, 1))
}
