% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means_by_muscle.R
\name{means_by_muscle}
\alias{means_by_muscle}
\title{Means by Muscle}
\usage{
means_by_muscle(df_no_means)
}
\arguments{
\item{df_no_means}{\code{data.frame} with two rows for each muscle}
}
\value{
\code{data.frame} with means aggregated for each muscle.
}
\description{
Aggregate means for each muscle.
}
\examples{
\dontrun{
df <- data.frame(
  muscle = c("mPTd", "mPTd"),
  x = c(30, 40), y = c(35, 45), z = c(35, 15)
)
means_by_muscle(df)
}
}
