% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_coords.R
\name{merge_coords}
\alias{merge_coords}
\title{Merge coordinates files}
\usage{
merge_coords(M1, M2)
}
\arguments{
\item{M1}{`data.frame` with coordinate data.}

\item{M2}{`data.frame` with coordinate data.}
}
\value{
`data.frame` with merged coordinate data.
}
\description{
`merge_coords()` takes two coordinates files and merges them into
one file. Variable names are post-pended with '_1' and '_2'.
}
\examples{

AL_008 <- read_coords(system.file("extdata",
                                  "AL_008_coords.xlsx",
                                  package = "MuscleTernary"),
                      system.file("extdata",
                                  "AL_008_forces.xlsx",
                                  package = "MuscleTernary"),
                      L_R_means = TRUE)
AL_031 <- read_coords(system.file("extdata",
                                  "AL_031_coords.xlsx",
                                  package = "MuscleTernary"),
                      system.file("extdata",
                                  "AL_031_forces.xlsx",
                                  package = "MuscleTernary"),
                      L_R_means = TRUE)

M <- merge_coords(AL_031, AL_008)
M

}

