% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscle_color_map.R
\name{muscle_fill_map}
\alias{muscle_fill_map}
\title{Holliday Lab Muscle Color Map}
\usage{
muscle_fill_map()
}
\value{
A \code{scale_fill_manual()} that can be added to a
  ggplot
}
\description{
This function returns a \code{scale_fill_manual()} that matches
the Holliday lab muscle color map.
}
\examples{
muscle_fill_map()
}
