% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-MuscleTernary.R
\name{norm}
\alias{norm}
\title{Normalize a vector (L2 norm)}
\usage{
norm(u)
}
\arguments{
\item{u}{Vector}
}
\value{
L2 norm of `u`.
}
\description{
Normalize a vector (L2 norm)
}
\keyword{internal}
