% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coords.R
\name{process_coords}
\alias{process_coords}
\title{Process coordinates data.frame}
\usage{
process_coords(coords, grouping, forces = TRUE, L_R_means = TRUE)
}
\arguments{
\item{coords}{`data.frame` containing coordinates and other
columns to be plotted in ternary space}

\item{grouping}{Character vector of grouping variables}

\item{forces}{If `TRUE` (default), tehen check for a column of
force data.}

\item{L_R_means}{If `TRUE` (default), the means of left and right
muscle are returned.}
}
\value{
\code{data.frame} suitable for plotting with \code{ggtern()}.
}
\description{
Process coordinates data.frame
}
