% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stl.R
\name{read_stl}
\alias{read_stl}
\title{Import STL mesh}
\usage{
read_stl(fname)
}
\arguments{
\item{fname}{string path to stl file}
}
\value{
numeric matrix with (x, y, z) coordinates of vertices on successful
read. `NA` on any warning.
}
\description{
This is a wrapper of `rgl::readSTL()`, which returns `NA` if the read
fails (which it does for both ASCII STLs and binary STLs that use big
endian encoding).
}
\examples{

read_stl(system.file("extdata",
                     "L_mAMEM_or.stl",
                     package = "MuscleTernary"))

}
