% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-MuscleTernary.R
\name{RU}
\alias{RU}
\title{Calculate rotation matrix}
\usage{
RU(A, B)
}
\arguments{
\item{A}{Vector (usually the unit vector)}

\item{B}{Vector}
}
\value{
U rotation matrix between A and B
}
\description{
Calculate the rotation matrix required to rotate vector A onto
vector B. Both vectors will be normalized to a unit vectors prior to
rotation.
See: http://math.stackexchange.com/a/897677/6965
}
\keyword{internal}
