% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-MuscleTernary.R
\name{ssc}
\alias{ssc}
\title{Skew-symmetric cross-product}
\usage{
ssc(v)
}
\arguments{
\item{v}{Matrix}
}
\value{
Skew-symmetric cross-product of `v`
}
\description{
Calculate the skew-symmetric cross-product matrix of v
See: See: http://math.stackexchange.com/a/897677/6965
}
\keyword{internal}
