% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{theta}
\alias{theta}
\title{Compute 3D vector cross product}
\usage{
theta(vector1, vector2)
}
\arguments{
\item{vector1}{numeric vector of length 3}

\item{vector2}{numeric vector of length 3}
}
\value{
Angle 'theta' between the two input vectors in radians
}
\description{
Calculates the angle between two vectors
}
\keyword{internal}
