% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_arrows.R
\name{write_arrows}
\alias{write_arrows}
\title{Write Maya arrows code}
\usage{
write_arrows(
  muscle,
  side,
  x_origin,
  y_origin,
  z_origin,
  x_insertion,
  y_insertion,
  z_insertion,
  force,
  cylinder_r,
  cone_r,
  cone_hr,
  outfile,
  rev_arrows
)
}
\arguments{
\item{muscle}{Muscle}

\item{side}{Side (L/R)}

\item{x_origin}{x origin}

\item{y_origin}{y origin}

\item{z_origin}{z origin}

\item{x_insertion}{x insertion}

\item{y_insertion}{y insertion}

\item{z_insertion}{z insertion}

\item{force}{Force (N)}

\item{cylinder_r}{Cylinder radius}

\item{cone_r}{Cone radius}

\item{cone_hr}{Cone height}

\item{outfile}{String file to write to}

\item{rev_arrows}{Boolean Reverse arrows or not}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of
  writing MEL commands to a file.
}
\description{
Write Maya arrows code
}
\examples{
\dontrun{
write_arrows(
  muscle = "mPTd", side = "L",
  x_origin = 10, y_origin = 20, z_origin = 30,
  x_insertion = 5, y_insertion = 15, z_insertion = 25,
  force = 100, cylinder_r = 2, cone_r = 4, cone_hr = 2,
  outfile = tempfile(fileext = ".mel"), rev_arrows = TRUE
)
}
}
