% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfiber_to_maya.R
\name{xfiber_to_maya}
\alias{xfiber_to_maya}
\title{Generate Maya mel file from xfiber tracing}
\usage{
xfiber_to_maya(fname, outfile, radius = 8, n = NULL)
}
\arguments{
\item{fname}{string: Path to excel file}

\item{outfile}{string: Path to output file}

\item{radius}{numeric: Radius of the cylinders in Maya}

\item{n}{numeric: Number of tracks to randomly select}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of
  writing a Maya mel script file.
}
\description{
Generate Maya mel file from xfiber tracing
}
\examples{
\dontrun{
f <- system.file("extdata", "AV069_SC.xml",
                 package = "MuscleTernary")
xfiber_to_maya(f, outfile = tempfile(fileext = ".mel"))
}
}
