% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprod.R
\name{xprod}
\alias{xprod}
\title{Cross-product of two 3d vectors}
\usage{
xprod(a, b)
}
\arguments{
\item{a}{numeric vector of length 3}

\item{b}{numeric vector of length 3}
}
\value{
numeric vector of length 3
}
\description{
This function calculates the cross-product of two 3d (3 element) vectors.
Note that this is different from `base::crossprod()`.
}
\examples{
xprod(c(0, 1, 2), c(1, 1, 1))

}
