% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_size.R
\name{centroid_size}
\alias{centroid_size}
\title{Calculate centroid size for an stl file}
\usage{
centroid_size(stl)
}
\arguments{
\item{stl}{String path to stl file}
}
\value{
Numeric Centroid size
}
\description{
Calculate centroid size for an stl file
}
