% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_ternary.R
\name{coords_to_ternary}
\alias{coords_to_ternary}
\title{Process coordinates data.frame}
\usage{
coords_to_ternary(coords, grouping = NULL)
}
\arguments{
\item{coords}{`data.frame` or `tibble` containing coordinates and other
columns to be plotted in ternary space. Required columns are at least
"muscle", "x_origin", "y_origin", "z_origin", "x_insertion",
"y_insertion", and "z_insertion"}

\item{grouping}{Character vector of grouping variables for calculating
means. Defaults to `NULL`, which will return all the original rows and
columns}
}
\value{
\code{data.frame} suitable for plotting with \code{ggtern()}.
}
\description{
Process coordinates data.frame
}
