% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ends_to_vectors.R
\name{ends_to_vectors}
\alias{ends_to_vectors}
\title{Function turning xfiber endpoint vectors into unit vectors
Should be used following 'find_track_ends'}
\usage{
ends_to_vectors(coords, grouping = NULL)
}
\arguments{
\item{coords:}{`data.frame` or `tibble` containing coordinates and other
columns to be converted into unit vectors. Required columns are at least
"muscle", "x_origin", "y_origin", "z_origin", "x_insertion",
"y_insertion", and "z_insertion".}

\item{grouping:}{Character vector of grouping variables for calculating
means. Defaults to `NULL`, which will return all the original rows and
columns}
}
\value{
'data.frame' with x, y, and z components of unit vectors of each
fiber track
}
\description{
Function turning xfiber endpoint vectors into unit vectors
Should be used following 'find_track_ends'
}
\examples{
D <- read_xfiber_xml(system.file("extdata",
                     "AV069_SC.xml",
                     package = "MuscleTernary")) |>
                     mutate(muscle = "SC")
ends <- find_track_ends(D)

ends_to_vectors(ends)


}
