% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_track_ends.R
\name{find_track_ends}
\alias{find_track_ends}
\title{Reduce Xfiber tracks to vectors connecting the two endpoints}
\usage{
find_track_ends(Tracks)
}
\arguments{
\item{Tracks}{data.frame: Object containing the full set of tracks and a
column for `muscle`, which is required by `coords_to_ternary()`.}
}
\value{
tibble: `Tracks` but reduced to one row per track.
}
\description{
This function finds the (x, y, z) coordinates of the starting and endsing
points of an Xfiber tracks object read by `read_xfiber_xml()`.
}
