% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_shiftedtext.R
\name{geom_shiftedtext}
\alias{geom_shiftedtext}
\title{geom for labeling geom_points}
\usage{
geom_shiftedtext(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  parse = FALSE,
  ...
)
}
\arguments{
\item{mapping}{The aes mapping}

\item{data}{The data}

\item{stat}{The stat}

\item{position}{The position}

\item{parse}{Parse or not}

\item{...}{Additional options passed on}
}
\value{
A geom usable by ggplot()
}
\description{
See: http://stackoverflow.com/q/19694497/168137
}
\examples{
\dontrun{
geom_shiftedtext(ggplot2::aes(label = muscle))
}
}
