% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mel.R
\name{make_mel}
\alias{make_mel}
\title{Write mel file}
\usage{
make_mel(stl, data, scale_radius = TRUE, max_radius = 8, rev_arrows = TRUE)
}
\arguments{
\item{stl}{String with the stl file. Assumed to present in the working
directory}

\item{data}{Data.frame with the data}

\item{scale_radius}{Boolean (default TRUE) Should the radius be scaled to
the maximum value}

\item{max_radius}{Numeric Maximum radius value}

\item{rev_arrows}{Boolean Should the arrows be reversed (probably TRUE)}
}
\description{
Write mel file
}
