% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mel.R
\name{make_mel}
\alias{make_mel}
\title{Write mel file}
\usage{
make_mel(
  stl,
  data,
  shader_file = "default",
  outfile = NULL,
  scale_radius = TRUE,
  max_radius = 8,
  rev_arrows = TRUE
)
}
\arguments{
\item{stl}{string Location of the stl file. Assumed to present in the
working directory}

\item{data}{data.frame Object with the data}

\item{shader_file}{string Either "default" to use the standard palette,
or the path to a csv file with colors for each muscle.}

\item{outfile}{string Name of file to write to. Defaults to the base name
of the stl file.}

\item{scale_radius}{boolean (default \code{TRUE}) Should the radius be
scaled to the maximum value automatically based on the stl centroid size?}

\item{max_radius}{numeric Maximum radius value. Ignored if
\code{scale_radius} is \code{TRUE}.}

\item{rev_arrows}{boolean (default \code{TRUE}) Should the arrowheads be
reversed (you probably want \code{TRUE})}
}
\description{
Write mel file
}
