% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscle_force.R
\name{muscle_force}
\alias{muscle_force}
\title{Estimate muscle force}
\usage{
muscle_force(PCSA, Tspec = 0.3)
}
\arguments{
\item{PCSA}{numeric: Value for PCSA (e.g., from \code{pcsa()})}

\item{Tspec}{numeric: Value for specific tension of muscle in cubic mm.
Default of 0.3 is reasonable for mammalian muscle.}
}
\value{
numeric: Estimate of muscle force
}
\description{
Use PCSA and specific tension to estimate muscle force
}
