% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Muscle_Force.R
\name{Muscle_Force}
\alias{Muscle_Force}
\title{Estimate muscle force}
\usage{
Muscle_Force(PCSA, Tspec = 0.3)
}
\arguments{
\item{PCSA}{numeric: Value for PCSA (e.g., from \code{PCSA()})}

\item{Tspec}{numeric: Value for specific tension of muscle in cubic mm.}
}
\value{
numeric: Estimate of muscle force
}
\description{
Use PCSA and specific tension to estimate muscle force
}
