% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuscleTernary-package.R
\docType{package}
\name{MuscleTernary-package}
\alias{MuscleTernary-package}
\alias{MuscleTernary}
\title{Muscle Ternary}
\description{
More description
}
\details{
\tabular{ll}{ Package: \tab MuscleTernary\cr
              Type: \tab Package\cr
              Version: \tab 0.25\cr
              Date: \tab 2025-11-11\cr
              License: \tab GPL\cr
              LazyLoad: \tab yes\cr
              LazyData: \tab yes\cr }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Middleton-Lab/MuscleTernary}
  \item Report bugs at \url{https://github.com/Middleton-Lab/MuscleTernary/issues}
}

}
\author{
Kevin M. Middleton (\url{middletonk@missouri.edu})
}
\keyword{internal}
\keyword{package}
