% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuscleTernary-package.R
\docType{package}
\name{MuscleTernary-package}
\alias{MuscleTernary-package}
\alias{_PACKAGE}
\alias{MuscleTernary}
\title{Muscle Ternary}
\description{
More description
}
\details{
\tabular{ll}{ Package: \tab MuscleTernary\cr
              Type: \tab Package\cr
              Version: \tab 0.24\cr
              Date: \tab 2023-03-17\cr
              License: \tab GPL\cr
              LazyLoad: \tab yes\cr
              LazyData: \tab yes\cr }
}
\author{
Kevin M. Middleton (\url{middletonk@missouri.edu})
}
\keyword{internal}
\keyword{package}
