% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCSA.R
\name{PCSA}
\alias{PCSA}
\title{Estimate PCSA from from a pair of muscle attachment stls.}
\usage{
PCSA(stl1, stl2, fascicle_length = 1, theta = 0, units_adjust = 1)
}
\arguments{
\item{stl1}{string: Path to stl.}

\item{stl2}{string: Path to stl.}

\item{fascicle_length}{numeric: Fascicle length}

\item{theta}{numeric: Fascicle angle (radians)}

\item{units_adjust}{numeric: Multiplier adjustment if units are not in mm.}
}
\value{
numeric: Estimate of PCSA for the muscle defined by the two
attachments of `stl1` and `stl2`.
}
\description{
Estimate PCSA from from a pair of muscle attachment stls.
}
