% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pennation_angle.R
\name{pennation_angle}
\alias{pennation_angle}
\title{Find pennation angle}
\usage{
pennation_angle(vector_table, central_axis)
}
\arguments{
\item{vector_table:}{'data.frame' or 'tibble 'with unit vectors produced by
'ends_to_vectors()'}

\item{central}{axis: Named numeric vector such as that produced by
'central_axis()'}
}
\value{
Xfiber dataframe/tibble with newly appended "PennationAngle" column
}
\description{
Measures pennation angles by calculating the angle
between the central axis and each fiber endpoint vector
}
