% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_coords.R
\name{read_coords}
\alias{read_coords}
\title{Read and process coordinates file}
\usage{
read_coords(coords_file, force_file = NULL, force_factor = 1)
}
\arguments{
\item{coords_file}{String with path to coordinates file.}

\item{force_file}{(Optional) String with path to force file.}

\item{force_factor}{Scaling factor for force. Defaults to 1.}
}
\value{
\code{data.frame} suitable for plotting with \code{ggtern()}.
}
\description{
Read and process coordinates file
}

