% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xfiber_xml.R
\name{read_xfiber_xml}
\alias{read_xfiber_xml}
\title{Import from Avizo xfiber XML}
\usage{
read_xfiber_xml(filename)
}
\arguments{
\item{filename}{string: file name to load (should be an XML file).}
}
\value{
tibble: All tracks concatenated.
}
\description{
Import from Avizo xfiber XML
}
