% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relative_proportion.R
\name{relative_proportion}
\alias{relative_proportion}
\title{Calculate the relative proportion of the items of a given vector}
\usage{
relative_proportion(x)
}
\arguments{
\item{x}{Vector of length 3}
}
\value{
A normalized unit vector of length 3
}
\description{
Calculate the relative proportion of the items of a given vector
}

