% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_area.R
\name{stl_area}
\alias{stl_area}
\title{Calculate area of an stl mesh}
\usage{
stl_area(fname)
}
\arguments{
\item{fname}{String: path to stl file}
}
\value{
Numeric: Area of the stl mesh
}
\description{
Calculate area of an stl mesh
}
