% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_arrows.R
\name{write_arrows}
\alias{write_arrows}
\title{Write Maya arrows code}
\usage{
write_arrows(
  muscle,
  side,
  x_origin,
  y_origin,
  z_origin,
  x_insertion,
  y_insertion,
  z_insertion,
  force,
  cylinder_r,
  cone_r,
  cone_hr,
  outfile,
  rev_arrows
)
}
\arguments{
\item{muscle}{Muscle}

\item{side}{Side (L/R)}

\item{x_origin}{x origin}

\item{y_origin}{y origin}

\item{z_origin}{z origin}

\item{x_insertion}{x insertion}

\item{y_insertion}{y insertion}

\item{z_insertion}{z insertion}

\item{force}{Force (N)}

\item{cylinder_r}{Cylinder radius}

\item{cone_r}{Cone radius}

\item{cone_hr}{Cone height}

\item{outfile}{String file to write to}

\item{rev_arrows}{Boolean Reverse arrows or not

 @export}
}
\description{
Write Maya arrows code
}
